Visual / Text tweaks (#194)
* Tweaks to the navbar text to make some of the headings fit on one line.

Make the final PIA number very visible and easy to find on the page.
User is often copying this number into another tool, or even this tool
for use as a spousal value. Making it easier to find improves usability.

* Make the text clearer that early / delayed adjustments apply on a
monthly basis, not on an annual basis. The text was ambiguous on this.

* Adjust bolding of a character.