RecipientName component. (#244)
This allows us to insert the name of the recipient into the report
in different places. The recipient will also be styled differently
if it is the first user or the second user.

If there is only one user, then the name will be presented as "you",
"your", etc as defined by the report. These variants are not styled.

In this way, the reports are clear as to which user is being described.