Reduce the amount of color.
With two users, the colors on names gets a little busy.

This change removes the username headers, instead folding them
into the heading for the "Earnings Record", and removing the color.

This also removes the green color from the big "Eligible" text, though
the red is still maintained for ineligibile users.

Overall, this makes the page a little less busy, and a little more
readable.