Merge svelte changes into main branch (#259)
* Initial svelte setup. (#220)

- Removed most of the angular code.
- Set up front page and calculator app with just some CSS for the paste prompt.

* Converting chunks of the logic to typescript (#222)

* Initial svelte setup.

- Removed most of the angular code.
- Set up front page and calculator app with just some CSS for the paste prompt.

* Rename `lib` to `components`

* rename `lib` to components

* Add storybook

* stage unstaged

* Set a 320px global min-width

* Paste Prompt story and CSS tweaks

* Start working on some of the logic libraries.

This covers constants, birthday, earnings-record, most of recipient,
and ssa-parse.

* PastePrompt / PasteConfirm

* Initial svelte setup.

- Removed most of the angular code.
- Set up front page and calculator app with just some CSS for the paste prompt.

* Rename `lib` to `components`

* rename `lib` to components

* Add storybook

* stage unstaged

* Set a 320px global min-width

* Paste Prompt story and CSS tweaks

* Start working on some of the logic libraries.

This covers constants, birthday, earnings-record, most of recipient,
and ssa-parse.

* PastePrompt and PasteConfirm components.

Moved these to a super-component of PasteFlow to simplify the
App level code.

* Added Paste Apology to the Paste Flow. (#224)

* Svelte work (#225)

* Added Paste Apology to the Paste Flow.

* Add AgeRequest to PasteFlow.

* Svelte work (#226)

* Added Paste Apology to the Paste Flow.

* Add AgeRequest to PasteFlow.

* Create a <div> to contain PasteFlow.

* Adding PasteFlow Story (#227)

* PasteFlow story, identified / fixed small issues.

* Update stories to use some organization hierarchy

* tiny cleanup

* Add a subscribe method to the recipient class. (#228)

This method follows the Svelte store contract, and returns an
unsubscribe function.

Also updated minor comments in PasteFlow.svelte.

* Added a Slider component to support the EarningsTable. (#229)

EarningsTable in progress, but incomplete.

* Earnings Record Report added to the App (#230)

* Implementing the Earnings Record Flow.

* cleanup

* Rename EarningsFlow to EarningsReport. (#231)

* Add EligibilityReport. (#232)

Supporting work adding EligibilityTable and Expando components.

Minor changes to some of the earnings report components to match
the new EligibilityReport's interface.

* Add a sidebar to the report. (#233)

Still to be implemented is an IntersectionObserver based system to
highlight the current section in the sidebar.

* Implement a sidebar component. (#234)

* Implement a sidebar component.

The sidebar uses IntersectionObserver to detect when each section is
visible, and highlights the corresponding section in the sidebar.

* Fix css.

* Create a new "Money" abstraction that holds currency amounts in cents. (#236)

Use this throughout the app instead of raw numbers to avoid floating
point rounding errors.

* Add chevron to sidebar. (#237)

* PIA report. (#238)

Still missing the adjustable canvas chart.

* Add the BendpointChart component. (#239)

* Add NormalRetirementAgeReport (#240)

* fix issue with PIA label

* NormalRetirementAgeReport component.

* Add FilingDateReport component, minus the canvas chart. (#241)

* Filing Date Chart using Svelte (#242)

* Implement a new Filing Date Chart, modelled after the spousal
chart from the previous version, but for only a single recipient.

* Implement a FilingDate Chart. Based on the Spousal Chart, but for
a single recipient.

* Modify the input flow to allow for choosing to enter data for (#243)

the user's spouse, in addition to themselves. Update the
report to show both users.

* RecipientName component. (#244)

This allows us to insert the name of the recipient into the report
in different places. The recipient will also be styled differently
if it is the first user or the second user.

If there is only one user, then the name will be presented as "you",
"your", etc as defined by the report. These variants are not styled.

In this way, the reports are clear as to which user is being described.

* Reduce the amount of color. (#245)

With two users, the colors on names gets a little busy.

This change removes the username headers, instead folding them
into the heading for the "Earnings Record", and removing the color.

This also removes the green color from the big "Eligible" text, though
the red is still maintained for ineligibile users.

Overall, this makes the page a little less busy, and a little more
readable.

* Spousal Section, including combined filing date chart. (#246)

* Get things working in preview mode. (#247)

* Get things working in preview mode.

* Remove analytics.

* Clean up positioning of sticky elements from the future (#248)

earnings sliders.

* Responsive focus changes. (#249)

Attempt to make all of the components work down to small mobile screen
widths. The combined chart still doesn't fit on the smallest of screens
but now there is a warning suggesting the user rotate the device.

* Add a more resources section. (#250)

* Small cleanup (#251)

* Add static files.

* Remove unused css

* Add alt tags. (#252)

* Sveltekit (#253)

* Converting to SvelteKit

* Migrate to sveltekit

* Clean up broken links. (#254)

* Clean up broken links.

* Fix another link.

* another fix

* Small tweaks to text

* Various cleanups from feedback (#255)

* Affect/Effect, reflect updates in earnings report

* Fix AIME acronym

* Fix some imports

* Fix typo, missing "T".

* Fix missing references to stores.

* Fix some broken links

* Make the expando transition optional as a hack

* Create an earnings report section at the top.

* Ensure that the filing date chart doesn't go wider (#256)

than the future sliders in a wide screen situation.

* Fix issues with pointer interactions. (#257)

Charts and Sliders were not working properly for move events triggered
on a touchscreen device. This should resolve that issue.

* Various fixes (#258)

* Fix issues with pointer interactions.

Charts and Sliders were not working properly for move events triggered
on a touchscreen device. This should resolve that issue.

* Clarity copy/paste text on mobile devices.

* Random fixes.