Use the recipient store less frequenct.
Now that I better understand the `$` syntax's behavior, I can see that
in a number of places, we are subscribing to updates when they are
not relevant. Most of these are for the recipient names, which don't
change after the intial flow.

This change reduces the number of subscriptions, which should have the
effect of reducing some cpu from callbacks. It's probably minimal, but
at least it's a little cleaner.