Fixes #306 (#309)
If we know the maximum earnings for a future year, which happens late
in the current year, allow the future wage slider to go up to that new
value.

This also enforces capping future earnings wages to the maximum earnings
for that year, if present. The result is that even if a user sets the
slider to a value higher than the maximum earnings for the current year,
the value will still be capped to the maximum earnings for that year.